Fixed some bugs relating to sending password reset link transporter
